==> code/05_word_embedding.Rmd <==
---
title: "04_word_embedding"
author: "Jae Yeon Kim"
date: "`r Sys.Date()`"
output: html_document
---

# Import libs and files 

## Libs

```{r}

# Import pkgs
pacman::p_load(
  tidyverse,
  lubridate, 
  purrr,
  furrr,
  tm,
  quanteda, 
  here,
  tidytext, # text analysis 
  textclean,
  textstem, 
  ggthemes,
  text2vec, # word embedding 
  widyr,
  patchwork, # arranging ggplots
  glue,
  # deep learning 
  text2vec,
  # network graph
  igraph,
  ggraph,
  # structural breakpoint
  strucchange,
  pbapply
)

devtools::install_github("prodriguezsosa/conText")
library(conText)
theme_set(theme_clean())

source(here("functions", "utils.R"))
```

## Files 

```{r}
load(here("processed_data", "df_ready.Rdata"))
```

```{r}
df_usa$org_label <- df_usa$Label 
corpus <- df_usa %>%
  filter(Label %in% c("Hate", "Counterhate")) %>%
  mutate(Label = if_else(Label == "Hate", 1, 0))
rm(df_usa)
```

```{r}
corpus$date <- as.POSIXct(corpus[["created_at"]], tz = "UTC", format = "%a %b %d %H:%M:%S %z %Y") %>% word(1)

corpus <- corpus %>%
  mutate(trump_speech = if_else(date >= as.Date(c("2020-03-16")), 1, 0)) %>%
  select(-c("created_at"))

corpus$trump_speech[1:5]
```

```{r}
corpus$date <- as.Date(corpus$date)
corpus$clean_text <- clean_text(corpus$full_text)

test <- corpus %>%
  group_by(trump_speech) %>%
  mutate(hate = if_else(org_label == "Hate", 1, 0),
         counterhate = if_else(org_label == "Counterhate", 1, 0)) 

test %>%
  summarize(hate_mean = mean(hate),
            counterhate_mean = mean(counterhate))

table(corpus$Label, 
      corpus$trump_speech)

t.test(subset(test$hate, test$trump_speech == 0),
       subset(test$hate, test$trump_speech == 1)) 

t.test(subset(test$counterhate, test$trump_speech == 0),
       subset(test$counterhate, test$trump_speech == 1)) 
```

# Word embedding 

## Local embedding and matrix 

```{r}
set.seed(20200921)
local_glove <- df2vec(corpus)
local_transform <- df2ltm(corpus, local_glove)

save(local_glove, local_transform, 
     file = here("processed_data/context_bg.Rdata"))
```

## ALC embedding approach 

### Regression 

```{r}
load(here("processed_data/context_bg.Rdata"))

mod <- conText(formula = chinesevirus ~ trump_speech + Label, 
          data = corpus, 
          text_var = 'clean_text', 
          pre_trained = local_glove,
          transform = TRUE, 
          transform_matrix = local_transform, 
          bootstrap = TRUE, 
          num_bootstraps = 20, 
          stratify_by = c('trump_speech', 'Label'),
          permute = TRUE, 
          num_permutations = 100, 
          window = 6, 
          valuetype = 'fixed', 
          case_insensitive = TRUE, 
          hard_cut = FALSE, 
          verbose = FALSE)

plot_tibble <- mod$normed_betas %>% 
  mutate(Coefficient = c("Trump tweet", "Hate label")) %>% 
  mutate(Coefficient = factor(Coefficient, levels = Coefficient))

plot_tibble
```

```{r}
reg_plot <- plot_tibble %>%
  ggplot(aes(x = Coefficient, y = Normed_Estimate)) + 
  geom_bar(position = position_dodge(), 
    stat = "identity", width = 0.5) + 
  geom_errorbar(aes(ymin = Normed_Estimate - 
    1.96 * Std.Error, ymax = Normed_Estimate + 1.96 * Std.Error), size = 0.75, width = 0.15, 
    position = position_dodge(0.9)) + 
  ylab("Norm of beta hats") + 
  theme(axis.text.x = element_text(size = 10, angle = 90, vjust = 0.5, hjust = 1),
          axis.text.y = element_text(size = 10)) +
  labs(title = "Keyword: Chinese virus")

reg_plot
ggsave(here("outputs", "embedreg.png"))
ggsave(here("outputs", "embedreg.pdf"))
```

# Exploration 

```{r}
corpus$full_text[str_detect(corpus$full_text, "selling")][1:3]
```

## Nearest neighbors 

### ALC embedding 

```{r}
set.seed(1234L)
hate_post_bt <- get_bt_terms(1, 1, "chinesevirus", 100)
hate_pre_bt <- get_bt_terms(0, 1, "chinesevirus", 100)
counterhate_post_bt <- get_bt_terms(1, 0, "chinesevirus", 100)
counterhate_pre_bt <- get_bt_terms(0, 0, "chinesevirus", 100)
```

```{r}
terms2plot(hate_pre_bt, counterhate_pre_bt, "Chinese virus", "Pre-Trump tweet") + terms2plot(hate_post_bt, counterhate_post_bt, "Chinese virus", "Post-Trump tweet")

ggsave(here("outputs", "btchinese.png"), width = 10, height = 10)
ggsave(here("outputs", "btchinese.pdf"), width = 10, height = 10)
```

# Additional tests 

```{r}
corpus_copy <- corpus %>%
  mutate(bat = if_else(str_detect(full_text, "bat"), 1, 0),
         weapon = if_else(str_detect(full_text, "weapon"), 1, 0),
         propaganda = if_else(str_detect(full_text, "propaganda"), 1, 0)) %>%
  mutate(trump_speech = recode(trump_speech, 
                               `1` = "Post-Trump tweet",
                               `0` = "Pre-Trump tweet")) %>%
  mutate(trump_speech = factor(trump_speech, 
                               levels = c("Pre-Trump tweet", "Post-Trump tweet")),
         Label = recode(Label, 
                        `1` = "Hate speech",
                        `0` = "Counterspeech"))

corpus_copy %>%
  group_by(Label) %>%
  summarize(mean_bat = round(mean(bat),3),
            mean_propaganda = round(mean(propaganda),3))

bat_sum <- corpus_copy %>%
  group_by(trump_speech, Label) %>%
  summarize(mean = mean(bat),
            ci_high = ci.high(bat),
            ci_low = ci.low(bat))

weapon_sum <- corpus_copy %>%
  group_by(trump_speech, Label) %>%
  summarize(mean = mean(weapon),
            ci_high = ci.high(weapon),
            ci_low = ci.low(weapon))

propaganda_sum <- corpus_copy %>%
  group_by(trump_speech, Label) %>%
  summarize(mean = mean(propaganda),
            ci_high = ci.high(propaganda),
            ci_low = ci.low(propaganda))

bat_propaganda_sum <- bind_rows(
  mutate(bat_sum, Keyword = "Bat"),
  mutate(weapon_sum, Keyword = "Weapon"),
  mutate(propaganda_sum, Keyword = "Propaganda"))
```

```{r}
bat_propaganda_sum %>%
  mutate(Keyword = factor(Keyword, levels = c("Bat", "Weapon", "Propaganda"))) %>%
  ggplot(aes(x = trump_speech, y = mean, 
             ymax = mean + ci_high, 
             ymin = mean - ci_low)) + 
  geom_pointrange() +
  facet_grid(Keyword~Label) +
  ggrepel::geom_text_repel(aes(label = glue("{round(mean, 3)*100}%"))) +
  scale_y_continuous(labels = scales::percent) +
  labs(x = "", 
       y = "Percentage of tweets")

ggsave(here("outputs", "batpropaganda.png"))

2.65/0.52
```

# Inspection

```{r}
corpus$trump_speech[corpus$trump_speech == 1] <- "Post-Trump tweet"

corpus$trump_speech[corpus$trump_speech == 0] <- "Pre-Trump tweet"

corpus$trump_speech <- factor(corpus$trump_speech, levels = c("Pre-Trump tweet", "Post-Trump tweet"))
```

```{r}
corpus %>%
  group_by(Label, trump_speech) %>%
  mutate(hate = if_else(Label == 1, 1, 0),
         counterhate = if_else(Label != 1, 1, 0)) %>%
  summarize("Hate" = sum(hate),
            "Counterhate" = sum(counterhate)) %>% 
  pivot_longer(c("Hate", "Counterhate"),
               "type") %>%
  ggplot(aes(x = trump_speech, y = value, fill = type)) +
  geom_col(position = position_dodge()) +
  labs(fill = "Label",
       y = "Count of Tweets",
       x = "")

ggsave(here("outputs", "propchange.png"))
```

```{r}
subset(corpus, trump_speech == 0)$full_text[str_detect(subset(corpus, trump_speech == 0)$full_text, "machine")]

subset(corpus, trump_speech == 1)$full_text[str_detect(subset(corpus, trump_speech == 1)$full_text, "racism")]

subset(corpus, trump_speech == 1)$full_text[str_detect(subset(corpus, trump_speech == 1)$full_text, "sympathizer")]

```
